refactor: updated useHooks directory name to hooks
